Added Nearly Finished ContentList
Technically first version of ContentList, uploaded primarily to observe progress and receive feedback, unfinished and untested, but a majority of the intended functionality has been included. Note that changes to the content list are made in the class itself (providing menus and receiving user input), this logic can either remain or be largely exported to another class such as ContentManager or Controller depending on what would be the best way to design the application.