Working on the ContentFactory
Trying to allow the ContentFactory to create content and content lists. Have functionality in place for gathering most of the required info from the TMDB communicator, but I'll need a way to get the currently stored reviews from the database. I'll need access to the feedback factory since it's responsible for creating review objects.