Second Tiny Update to ContentList
Modified the removeContents method to sort the indexes to remove in reverse order to prevent issues with accessing the contents ArrayList outside its changing bounds. Also modified addContent (with another list of Content) so that it does not add the second list to the current one if the combined size would exceed the hard cap.